+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/query/plan/cascades/predicates/ConstantFoldingTest.java
@@ -20,231 +20,33 @@
 
 package com.apple.foundationdb.record.query.plan.cascades.predicates;
 
-import com.apple.foundationdb.record.Bindings;
-import com.apple.foundationdb.record.EvaluationContext;
-import com.apple.foundationdb.record.query.expressions.Comparisons;
-import com.apple.foundationdb.record.query.plan.cascades.AliasMap;
-import com.apple.foundationdb.record.query.plan.cascades.CorrelationIdentifier;
-import com.apple.foundationdb.record.query.plan.cascades.predicates.simplification.ConstantFoldingRuleSet;
-import com.apple.foundationdb.record.query.plan.cascades.typing.Type;
-import com.apple.foundationdb.record.query.plan.cascades.values.ConstantObjectValue;
-import com.apple.foundationdb.record.query.plan.cascades.values.LiteralValue;
-import com.apple.foundationdb.record.query.plan.cascades.values.NullValue;
-import com.apple.foundationdb.record.query.plan.cascades.values.Value;
-import com.apple.foundationdb.record.query.plan.cascades.values.simplification.Simplification;
-import com.google.common.base.Verify;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
+import com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.ValueWrapper;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Optional;
 import java.util.stream.Stream;
 
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.areEqual;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.areEqualAsRange;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.areNotEqual;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.covFalse;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.covNull;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.covTrue;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.isNotNull;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.isNotNullAsRange;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.isNull;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.isNullAsRange;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.litFalse;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.litNull;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.litTrue;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.notNullIntCov;
+import static com.apple.foundationdb.record.query.plan.cascades.ConstantFoldingTestUtils.simplify;
+
+@SuppressWarnings("PMD.TooManyStaticImports")
 public class ConstantFoldingTest {
-
-    private static int counter;
-
-    @SuppressWarnings("OptionalUsedAsFieldOrParameterType")
-    public static final class ValueWrapper {
-
-        @Nonnull
-        private final Optional<EvaluationContext> evaluationContext;
-
-        @Nonnull
-        private final Value value;
-
-
-        private ValueWrapper(@Nonnull final Value value, @Nonnull final Optional<EvaluationContext> evaluationContext) {
-            this.evaluationContext = evaluationContext;
-            this.value = value;
-        }
-
-        @Nonnull
-        public Value value() {
-            return value;
-        }
-
-        @Nonnull
-        public Optional<EvaluationContext> getEvaluationContextMaybe() {
-            return evaluationContext;
-        }
-
-        @Override
-        public String toString() {
-            if (value() instanceof ConstantObjectValue) {
-                Verify.verify(evaluationContext.isPresent());
-                return "(CoV → " + value().evalWithoutStore(evaluationContext.get()) + ")";
-            }
-            if (value instanceof LiteralValue<?>) {
-                return "(ℓ " + ((LiteralValue<?>)value).getLiteralValue() + ")";
-            }
-            Verify.verify(value instanceof NullValue);
-            return "∅";
-        }
-
-        @Nonnull
-        public EvaluationContext mergeEvaluationContext(@Nonnull final ValueWrapper that) {
-            if (evaluationContext.isEmpty() && that.evaluationContext.isEmpty()) {
-                return EvaluationContext.empty();
-            }
-            if (evaluationContext.isEmpty()) {
-                return that.evaluationContext.get();
-            }
-            if (that.evaluationContext.isEmpty()) {
-                return evaluationContext.get();
-            }
-            final var thisBindingsChildBuilder = evaluationContext.get().getBindings().childBuilder();
-            final var thatBindings = that.evaluationContext.get().getBindings();
-            thatBindings.asMappingList().forEach(entry -> thisBindingsChildBuilder.set(entry.getKey(), entry.getValue()));
-            return EvaluationContext.forBindings(thisBindingsChildBuilder.build());
-        }
-
-        @Nonnull
-        public static ValueWrapper of(@Nonnull final EvaluationContext evaluationContext, @Nonnull final ConstantObjectValue constantObjectValue) {
-            return new ValueWrapper(constantObjectValue, Optional.of(evaluationContext));
-        }
-
-        @Nonnull
-        public static ValueWrapper of(@Nonnull final Value value) {
-            return new ValueWrapper(value, Optional.empty());
-        }
-    }
-
-    @Nonnull
-    private static ValueWrapper newCov(@Nonnull final Type type, @Nullable Object bindingValue) {
-        final var correlationId = CorrelationIdentifier.uniqueID();
-        final var constantId = String.valueOf(counter++);
-        final var bindingKey = Bindings.Internal.CONSTANT.bindingName(correlationId.getId());
-        final var bindingValueMap = new HashMap<String, Object>();
-        bindingValueMap.put(constantId, bindingValue);
-        return ValueWrapper.of(EvaluationContext.forBinding(bindingKey, bindingValueMap), ConstantObjectValue.of(correlationId, constantId, type));
-    }
-
-    @Nonnull
-    static ValueWrapper litNull() {
-        return ValueWrapper.of(new NullValue(Type.nullType()));
-    }
-
-    @Nonnull
-    static ValueWrapper covNull() {
-        return newCov(Type.nullType(), null);
-    }
-
-    @Nonnull
-    static ValueWrapper litFalse() {
-        return ValueWrapper.of(LiteralValue.ofScalar(false));
-    }
-
-    @Nonnull
-    static ValueWrapper covFalse() {
-        return newCov(Type.primitiveType(Type.TypeCode.BOOLEAN), false);
-    }
-
-    @Nonnull
-    static ValueWrapper litTrue() {
-        return ValueWrapper.of(LiteralValue.ofScalar(true));
-    }
-
-    @Nonnull
-    static ValueWrapper covTrue() {
-        return newCov(Type.primitiveType(Type.TypeCode.BOOLEAN), true);
-    }
-
-    @Nonnull
-    static ValueWrapper notNullIntCov() {
-        return newCov(Type.primitiveType(Type.TypeCode.INT, false), 42);
-    }
-
-    @Nonnull
-    static RangeConstraints buildSingletonRange(@Nonnull Comparisons.Type comparisonType) {
-        return buildSingletonRange(comparisonType, null);
-    }
-
-    @Nonnull
-    static RangeConstraints buildSingletonRange(@Nonnull Comparisons.Type comparisonType, @Nullable final Value comparand) {
-        final var singletonRangeBuilder = RangeConstraints.newBuilder();
-        switch (comparisonType) {
-            case IS_NULL: // fallthrough
-            case NOT_NULL:
-                singletonRangeBuilder.addComparisonMaybe(new Comparisons.NullComparison(comparisonType));
-                break;
-            default:
-                singletonRangeBuilder.addComparisonMaybe(new Comparisons.ValueComparison(comparisonType, Verify.verifyNotNull(comparand)));
-        }
-        final var singletonRange = singletonRangeBuilder.build();
-        Assertions.assertThat(singletonRange).isPresent();
-        return singletonRange.get();
-    }
-
-    @Nonnull
-    static QueryPredicate isNotNull(@Nonnull final Value value) {
-        return new ValuePredicate(value, new Comparisons.NullComparison(Comparisons.Type.NOT_NULL));
-    }
-
-    @Nonnull
-    static QueryPredicate isNotNullAsRange(@Nonnull final Value value) {
-        return PredicateWithValueAndRanges.ofRanges(value, ImmutableSet.of(buildSingletonRange(Comparisons.Type.NOT_NULL)));
-    }
-
-    @Nonnull
-    static QueryPredicate isNull(@Nonnull final Value value) {
-        return new ValuePredicate(value, new Comparisons.NullComparison(Comparisons.Type.IS_NULL));
-    }
-
-    @Nonnull
-    static QueryPredicate isNullAsRange(@Nonnull final Value value) {
-        return PredicateWithValueAndRanges.ofRanges(value, ImmutableSet.of(buildSingletonRange(Comparisons.Type.IS_NULL)));
-
-    }
-
-    @Nonnull
-    static QueryPredicate areEqual(@Nonnull final Value value1, @Nonnull final Value value2) {
-        return new ValuePredicate(value1, new Comparisons.ValueComparison(Comparisons.Type.EQUALS, value2));
-    }
-
-    @Nonnull
-    static QueryPredicate areEqualAsRange(@Nonnull final Value value1, @Nonnull final Value value2) {
-        return PredicateWithValueAndRanges.ofRanges(value1, ImmutableSet.of(buildSingletonRange(Comparisons.Type.EQUALS, value2)));
-
-    }
-
-    @Nonnull
-    static QueryPredicate areNotEqual(@Nonnull final Value value1, @Nonnull final Value value2) {
-        return new ValuePredicate(value1, new Comparisons.ValueComparison(Comparisons.Type.NOT_EQUALS, value2));
-    }
-
-    @Nonnull
-    static QueryPredicate and(@Nonnull final QueryPredicate... preds) {
-        return AndPredicate.and(Arrays.stream(preds).collect(ImmutableList.toImmutableList()));
-    }
-
-    @Nonnull
-    static QueryPredicate or(@Nonnull final QueryPredicate... preds) {
-        return OrPredicate.or(Arrays.stream(preds).collect(ImmutableList.toImmutableList()));
-    }
-
-    @Nonnull
-    static QueryPredicate simplify(@Nonnull final QueryPredicate predicate) {
-        return simplify(predicate, EvaluationContext.empty());
-    }
-
-    @Nonnull
-    static QueryPredicate simplify(@Nonnull final QueryPredicate predicate,
-                                   @Nonnull final EvaluationContext evaluationContext) {
-        final var result = Simplification.optimize(predicate,
-                evaluationContext,
-                AliasMap.emptyMap(),
-                ImmutableSet.of(),
-                ConstantFoldingRuleSet.ofSimplificationRules());
-        return result.get();
-    }
 
     ///
     /// EQUALS simplification tests
